delete dd

+++ app/Http/Controllers/Api/SubTopicController.php
@@ -51,7 +51,7 @@ class SubTopicController extends ApiController
     {
         DB::enableQueryLog();
         try {
-            //$this->authorize('ticket.create');
+            $this->authorize('ticket.view');
             $pageSize    = $request->get('limit', 25);
             $this->trash = $this->trashStatus($request);
             $data        = $this->model->getByQuery($request->all(), $pageSize, $this->trash);
@@ -71,7 +71,7 @@ class SubTopicController extends ApiController
     public function show(Request $request, $id)
     {
         try {
-            //$this->authorize('ticket.view');
+            $this->authorize('ticket.view');
             $trashed = $request->has('trashed') ? true : false;
             $data    = $this->model->getById($id, $trashed);
             return $this->successResponse($data);
@@ -101,7 +101,7 @@ class SubTopicController extends ApiController
         DB::beginTransaction();
         DB::enableQueryLog();
         try {
-           // $this->authorize('ticket.create');
+            $this->authorize('ticket.create');
             $this->validate($request, $this->validationRules, $this->validationMessages);
             $model = $this->model->store($request->all());
             // dd(DB::getQueryLog());
@@ -146,7 +146,7 @@ class SubTopicController extends ApiController
         DB::beginTransaction();
         DB::enableQueryLog();
         try {
-            //$this->authorize('ticket.update');
+            $this->authorize('ticket.update');
             $this->validate($request, $this->validationRules, $this->validationMessages);
             $model = $this->model->update($id,$request->all());
             //dd(DB::getQueryLog());
@@ -191,7 +191,7 @@ class SubTopicController extends ApiController
         DB::beginTransaction();
         DB::enableQueryLog();
         try {
-            //$this->authorize('ticket.delete');
+            $this->authorize('ticket.delete');
             $this->model->delete($id);
             DB::commit();
             //dd(DB::getQueryLog());
